+++ packages/core/src/integrations/index.ts
@@ -4,3 +4,4 @@ export { SDKInformation } from './sdkinformation';
 export { InboundFilters } from './inboundfilters';
 
 export { Debug } from './pluggable/debug';
+export { RewriteFrames } from './pluggable/rewriteframes';